Relocate CreateDraft presenter + controller
