+++ tests/controllers/test_prefilled_draft_data_controller.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 from uuid import uuid4
 
 from arbeitszeit.use_cases.create_plan_draft import CreatePlanDraftRequest
-from arbeitszeit_web.create_draft import CreateDraftController
+from arbeitszeit_web.controllers.create_draft_controller import CreateDraftController
 from tests.forms import DraftForm
 from tests.session import FakeSession
