+++ tests/presenters/test_create_draft_presenter.py
@@ -2,7 +2,7 @@ from unittest import TestCase
 from uuid import uuid4
 
 from arbeitszeit.use_cases.create_plan_draft import CreatePlanDraftResponse
-from arbeitszeit_web.create_draft import CreateDraftPresenter
+from arbeitszeit_web.presenters.create_draft_presenter import CreateDraftPresenter
 from tests.presenters.notifier import NotifierTestImpl
 from tests.presenters.url_index import UrlIndexTestImpl
 from tests.translator import FakeTranslator
